mSecureHash & tests

+++ Code/util_sc/include_util_sc/crypto_util.h
@@ -1,7 +1,6 @@
 #ifndef __CRYPTO_UTIL_H
 #define __CRYPTO_UTIL_H
 
-#include <stddef.h>
 #include "defs_types.h"
 
 #ifdef __cplusplus
@@ -10,9 +9,10 @@ extern "C" {
 
 
 
-
+/********************************************************************/
 /******************************SHA256********************************/
 /********************************************************************/
+
 /****************************** MACROS ******************************/
 #define SHA256_BLOCK_SIZE 32            // SHA256 outputs a 32 byte digest
 
@@ -28,7 +28,7 @@ typedef struct {
 
 /*********************** FUNCTION DECLARATIONS **********************/
 void sha256_init(SHA256_CTX *ctx);
-void sha256_update(SHA256_CTX *ctx, const BYTE data[], size_t len);
+void sha256_update(SHA256_CTX *ctx, const BYTE data[], WORD len);
 void sha256_final(SHA256_CTX *ctx, BYTE hash[]);
